Clean up session ID:
The session ID was split between the generated ID and the handle,
which could serve as a unique address (not likely to be provided
more than once at the same process, right?). Now we're setting it
to only one of them.

Essentially, the session ID will simply be the address of the
client handle.