Implement remove_stream function:
I'm not really sure what this function does. The problem is that
there are multiple heaps and heap reference copies - global heap,
heap for each client, a reference to that in the global heap, and
of course the heap provided by POE. Difficult to fathom.

For now we're just implementing it by sort of copying it over,
and we'll debug this later.