Normalize key values for registrations:
We kept the registration with a leading underscore (such as
"_regex", "_subscribers", "_full", etc.), because they are
internals, but some methods ("status") require to seek out those
using a global variable (`@_STREAM_NAMES`, `@_STREAM_ASSISTANTS`),
and will thus break.

So, we normalize everything to what it was before. The next step
in this direction (which wouldn't be taken yet) will be moving
these from the keys to a heap for each Session ID.