Adding the client in one go:
This didn't take so long.

One thing that isn't supported yet and should be noted is the
ability to retrieve by block type. You have per line, hash, etc.

This uses a POE::Filter for it:

* POE::Filter::Stream - no-op
* POE::Filter::Line - per line

The block type changes the newline usage. I think the server
expects things to be per line, but it can tested using
AnyEvent::Handle's push_read( sub { ... } ) interface.

I'm not that inclined to add it, but meh.