Store the information of each client in their heap:
The POE client stores all the information in the HEAP, which is
in the main object ($self, for us). Instead, we're storing each
client's data (subscription, buffers, feed configuration, etc.)
in the client's heap ($self->{'clients'}{$SID}).

This allows us to do less overhead when creating and deleting
clients and when updating their information. The downside is that
we need to iterate over all clients when we need to perform an
action to know if we should do it for them or not.

A simple way to fix this in the future is maintain an index of
all registrations and configurations to session IDs. This is not
mandatory since I don't imagine the registrations to be massive,
but can't hurt (unless someone fiddles with it).

When it comes to buffers, we should possibly store the buffers
outside since they are not really per user. However, keeping them
in the client data allows us to have more flexible buffering.

This means that this change might be somewhat reverted (although
it is more likely to be complemented with a quick index and maybe
moving the buffers back to global namespace).

We also changed "subscribers" to "subscription" since it's a more
suitable name.

I don't like having no methods or attributes to any of this other
than "clients" but for now it's okay.

Stream assisters are still stored in the global namespace.