Make sure tests aren't accidentally skipped, fixes GH #2:
GH #2 notes that some tests might be skipped because a
`$cv->send` is often called when a line in an `else` clause, which
means that if a matching skipped (due to some error), the test
won't be run.

Since we're using `done_testing` instead of counting tests, we
won't know if that happens.

This commit fixes it by doing the following:

* Remove the `use_ok` checks from the t::lib::Eris::Test module.
* Add them to `t/server/basic.t`.
* (Add `t/client/basic.t` that does the same.)
* Add parameter of `plan` to t::lib::Eris::Test for convenience.
* Remove `done_testing` in favor of proper counting in all modules.

We also cleaned up the testing module helper.

On another note, AnyEvent::Log::fatal dies in an interesting way
which can't be detected with Test::Fatal. The docs for
AnyEvent::Log say you should redefine AnyEvent::Log::fatal_exit,
but that does not receive the error message, so that's useless.