Fix the successful connection callback to set up the pipe properly:
Serves me right for not having tests. Basically the handle object
was not running the 'on_connect' callback. I'm not sure why. Now
we're creating the handle and setting up the initial state and
creating the pipe timer.

I think that setting up those initial variables is silly, because
a boolean check on them would appear false anyway, so who cares
about their initial state if it's 0 or undef? Both false. Oh well.