Remove eris protocol and state from Client:
The eris protocol allows registering and unregistering. However,
the Client doesn't make use of unregistering.

The way it works is:

* A client provides in the constructor what it should connect to.
* When the client starts, it waits one second and then registers.
* The internal client message handler (*not* the user-provided
  "MessageHandler") has states ("readyState" and "connected").
* If it received a "EHLO" message, it marks "connected" as true.
* The following message will get checked against subcription
  responses. If they match, it marks "readyState" as true.
* All next requests are considered syslog messages which will
  be sent to the user-provided "MessageHandler".

This is pretty weird.

The only benefit that this check provides is handling unknown
messages, and it *only* does that *before* a subscription
confirmation message arrives. After that, it doesn't check anything
anymore, and you cannot downgrade any subscription. You can't even
issue a disconnect once the connection is confirmed ("readyState").

Another issue is that the subscription only happens a second later
("_setup_pipe_timer" timer and "setup_pipe" method), any messsages
queued right after the initialization will be ignored. This is
effectively a race condition, observed in a test.

This also depends on size of messages, since AnyEvent buffering is
quite elaborate, for better or worse.