Handle AnyEvent buffering better:
AnyEvent does smart buffering and the "on_read" callback might
receive multiple lines in a single call. We split it ourselves.

This is still not good enough because "on_read" callback will
be called without checking whether it finished an entire line,
because lines are arbitrary to this raw callback. This is by
design and perfectly fine, but we weren't counting on this so
this will be handled more intelligently as a single line input
in the next commit.