Move buffers outside the clients heap:
Until now we stored the buffers under the SID in the clients heap.

We now store all the buffers in the global heap (still under the
SID). Because we already have the buffers, this saves us to go
into the client heap and ask for the buffers.

Not the biggest save but why not.