Add benchmark scripts:
These benchmark scripts do the following:

* Measure how long 1K + 1 requests are done.
* Checking for memory leaks.

It does this by setting up a client and a server, and then sending
the server 1K requests + one additional sending request (sentinel).

Each test ignores the setup it can take and only measures the time
right before it sends the messages and as soon as it receives the
last message. Setup and teardown are not taken into account (as
they shouldn't be).

The test is run several times within a test module that measures
the memory allocation of the process and makes sure that it was
released once done.

This cannot check how many buffers were sent since this depends on
the event loop and underlying bookeeping that is not exposed to
the test.

There are three variables that can be changed in the code, at the
very top:

1. The amount of messages used in the test.
2. The amount of calls used by the memory checker module.
3. The amount of burn in calls used by the memory checker module.